+++ projects/maxDankow/src/test/java/ru/mipt/diht/students/maxDankow/sqlcollections/Statistics.java
@@ -1,0 +1,35 @@
+package ru.mipt.diht.students.maxDankow.sqlcollections;
+
+public class Statistics {
+
+    private final String group;
+    private final Long count;
+    private final Long age;
+
+    public String getGroup() {
+        return group;
+    }
+
+    public Long getCount() {
+        return count;
+    }
+
+    public Long getAge() {
+        return age;
+    }
+
+    public Statistics(String group, Long count, Long age) {
+        this.group = group;
+        this.count = count;
+        this.age = age;
+    }
+
+    @Override
+    public String toString() {
+        return "Statistics{"
+                + "group='" + group + '\''
+                + ", count=" + count
+                + ", age=" + age
+                + '}';
+    }
+}
